Fix examples type and add missing include
The first argument of `pok_thread_create()` is of type `uint32_t *`,
not `uint8_t *`.

+++ examples/mutexes/pr1/main.c
@@ -21,7 +21,7 @@
 uint8_t mid;
 
 int main() {
-  uint8_t tid;
+  uint32_t tid;
   pok_ret_t ret;
   pok_thread_attr_t tattr;
